pyproject.toml
[project]
name = "xtts-gtk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygobject>=3.52.3",
    "pyinstaller>=6.15.0",
    "pypinyin>=0.55.0",
]

[project.optional-dependencies]
tts = [
    "coqui-tts>=0.27.1",
    "coqpit-config>=0.2.1",
]

[tool.pyright]
typeCheckingMode = "basic"

[tool.setuptools]
py-modules = ["main", "settings", "tts_installer"]

[dependency-groups]
dev = [
    "pygobject-stubs>=2.13.0",
]
